fix: react native reanimated version and onboarding padding 🐛
