+++ package.json
@@ -43,7 +43,7 @@
     "react-native-gesture-handler": "^2.27.2",
     "react-native-gifted-charts": "^1.4.63",
     "react-native-linear-gradient": "^2.8.3",
-    "react-native-reanimated": "^3.18.0",
+    "react-native-reanimated": "^3.17.5",
     "react-native-safe-area-context": "^5.5.1",
     "react-native-simple-line-chart": "^0.37.0",
     "react-native-svg": "^15.12.0",
