+++ src/navigation/AppNavigator.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import { NavigationContainer } from '@react-navigation/native';
 import { createNativeStackNavigator } from '@react-navigation/native-stack';
 import { Box } from 'native-base';
-import { OnboardingFlow } from '../features/onboarding/screens/OnboardingFlow';
+import { OnboardingLayout } from '../features/onboarding/screens/OnboardingLayout';
 import { LoginScreen } from '../features/auth/screens/LoginScreen';
 import { RegisterScreen } from '../features/auth/screens/RegisterScreen';
 import HomeScreen from '../features/landing/screens/home';
@@ -64,7 +64,7 @@ export const AppNavigator = ({
           </Stack.Screen>
         ) : (
           <Stack.Screen name="Onboarding">
-            {() => <OnboardingFlow onComplete={handleOnboardingComplete} onLogin={handleLogin} />}
+            {() => <OnboardingLayout onComplete={handleOnboardingComplete} onLogin={handleLogin} />}
           </Stack.Screen>
         )
       ) : (
